qa: remove niche serializers to focus on the most popular serializers
This reduces the maintenance effort to a bare minimum.

`Wddx` extension is not available via sury or pecl. Compiling the extension from source does not work either and thus, we're dropping it.

`PythonPickle` seems to fill a very niche use-case where PHP and Python applications have to share data and thus, we are dropping support for it as well.

`MsgPack` requires an extension (the same way as `IgBinary` does). We are removing this serializer for now. In case the demand for it is overwhelming, we might bring it back.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>