qa: remove generic `Serializer` class
Static classes providing runtime features such as holding default serializers or generic factories are not supported by laminas anymore. If there is a strong need for this kind of logic, projects can write their own registry to persist these kinds of defaults.

Laminas components are primarily meant to be used in projects with dependency injection. Retrieving and replacing default serializers via the container configuration will always be an option.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>